apenas o usuario aluno pode entrar nas salas, as salas estao sendo criadas e sendo listadas tanto para os alunos quanto para os diretores

+++ S.A/src/main/java/com/sa/repository/InstituicaoRepository.java
@@ -17,6 +17,6 @@ public interface InstituicaoRepository extends JpaRepository<Instituicao, Long>{
 	public List<Instituicao> findByAlunosI(Aluno aluno);
 	public List<Instituicao> findByDiretor(Diretor diretor);
 	public List<Instituicao> findByProfessoresI(Professor professor);
-	public Instituicao findById(int id);
+	public Instituicao findById(long id);
 	
 }
